Moved functions to header file
